Move common fixtures to conftest.py
